TabBar: draw using natural colors.
The previous tab bar appearance was non-intuitive because the color scheme was inverted
from normal usage of tabs, whereby the tab item that is selected is normally
the same color as the page (thereby visually linking them).

This patch makes the tab item for the current page the same color as the form background,
and tab items for other pages are slightly darkened.

This improves the appearance considerably, IMHO.