Button.cpp: custom paint for non-pointer devices
For pointer devices (has_pointer() = true), this does nothing.
For non-pointer devices, draws buttons using the focus color used previously for
property highlighting.

This closes #495.

TabBar, XML: pass in form background color

This enables the form's color to be used for more natural looking tabs.

TabBar: draw using natural colors.

The previous tab bar appearance was non-intuitive because the color scheme was inverted
from normal usage of tabs, whereby the tab item that is selected is normally
the same color as the page (thereby visually linking them).

This patch makes the tab item for the current page the same color as the form background,
and tab items for other pages are slightly darkened.

This improves the appearance considerably, IMHO.

Form/List: use system-wide highlight color for selected item if focused.

This helps usability for non-pointer platforms.

Form/Screen: enable highlighting of focused items regardless of has_pointer()

Although for Altair we know we dont have a pointer, when users use other builds with
a pointer, the widget under focus should also be highlighted, as they may or may not
be using the keypad.

This also makes it quite clear when a dialog opens, what pressing enter will do.

System-wide colors

SDL/EditWindow